refactor: move saved alerts under settings for co-location/discoverability
