+++ src/v2/routes.tsx
@@ -35,7 +35,6 @@ import { paymentRoutes } from "v2/Apps/Payment/paymentRoutes"
 import { preferencesRoutes } from "./Apps/_Preferences2/preferencesRoutes"
 import { priceDatabaseRoutes } from "./Apps/PriceDatabase/priceDatabaseRoutes"
 import { purchaseRoutes } from "v2/Apps/Purchase/purchaseRoutes"
-import { savedSearchAlertsRoutes } from "v2/Apps/SavedSearchAlerts/savedSearchAlertsRoutes"
 import { searchRoutes } from "v2/Apps/Search/searchRoutes"
 import { settingsRoutes } from "v2/Apps/Settings/settingsRoutes"
 import { shippingRoutes } from "v2/Apps/Shipping/shippingRoutes"
@@ -83,7 +82,6 @@ export function getAppRoutes(): AppRouteConfig[] {
     { routes: preferencesRoutes },
     { routes: priceDatabaseRoutes },
     { routes: purchaseRoutes },
-    { routes: savedSearchAlertsRoutes },
     { routes: staticPageRoutes },
     { routes: searchRoutes },
     { routes: settingsRoutes },
